+++ packages/thirdweb/src/react/web/wallets/injected/locale/en.ts
@@ -16,7 +16,7 @@ const injectedWalletLocaleEn = (wallet: string): InjectedWalletLocale => ({
   scanScreen: {
     instruction: `Scan the QR code with the ${wallet} app to connect`,
   },
-  getStartedLink: `Don't have a ${wallet} wallet?`,
+  getStartedLink: `Don't have ${wallet}?`,
   download: {
     chrome: "Download Chrome Extension",
     android: "Download on Google Play",