+++ legacy_packages/auth/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @thirdweb-dev/auth
 
+## 4.1.87
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/wallets@2.5.29
+
+## 4.1.86
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @thirdweb-dev/wallets@2.5.28
+
 ## 4.1.85
 
 ### Patch Changes